refactored db config and added db/index.js, modified existing files accordingly
