+++ .env.example
@@ -1,0 +1,3 @@
+DATABASE_URL=
+PORT=
+API_KEY=
