+++ db/index.js
@@ -1,0 +1,4 @@
+const { Pool } = require('pg')
+const config = require('../config')
+
+module.exports = new Pool(config.db)
