+++ db/seed_database.js
@@ -1,4 +1,4 @@
-const db = require('../config')
+const db = require('../db')
 const fs = require('fs')
 
 let stations = fs.readFileSync('./resources/stations.csv', 'utf-8')
