+++ public/client.js
@@ -1,7 +1,5 @@
-let map = { lat: -33.8712, long: 151.2046 }
-
 async function initMap() {
-  map = new google.maps.Map(document.getElementById('map'), {
+  const map = new google.maps.Map(document.getElementById('map'), {
     center: { lat: -33.87995, lng: 151.200886 },
     zoom: 13,
     minZoom: 10,
@@ -16,7 +14,7 @@ async function initMap() {
   map.addListener('center_changed', () => {
     currentLocation.textContent = map.getCenter()
   })
-  
+
   fetch('/api/stations/all')
     .then((response) => response.json())
     .then((data) =>
