+++ public/css/style.css
@@ -5,7 +5,7 @@
 
 body {
     display: grid;
-    grid-template-columns: 1fr 600px 1fr;
+    grid-template-columns: 500px 1fr 500px;
 }
 
 #stats-table {
