+++ views/index.ejs
@@ -6,7 +6,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Project 3 Bugs Repellent</title>
     <link rel="stylesheet" href="./css/style.css">
-    
 </head>
 <body>
     <aside id="left">
@@ -50,7 +49,7 @@
             <h2>About</h2>
             <h3>Team Name: Bugs Repellent</h3>
             <p>Deepa <a href="" target="_blank">link here</a></p>
-            <p>Dimitri <a href="" target="_blank">link here</a></p>
+            <p>Dimitri <a href="https://github.com/dimitrinossar" target="_blank">https://github.com/dimitrinossar</a></p>
             <p>Kevin <a href="https://github.com/utamakevin" target="_blank">https://github.com/utamakevin</a></p>
             <p>Maria <a href="https://github.com/ValeGottardello" target="_blank">https://github.com/ValeGottardello</a></p>
             <p>&copy; Bugs Repellent</p>
@@ -79,7 +78,7 @@
     <script type="module" src="/components/clock.js"></script>
     <script type="module" src="/stations_api.js"></script>
     <script src="client.js"></script>
-    <script async src="https://maps.googleapis.com/maps/api/js?key=<%= API_KEY %>&callback=initMap"></script>
+    <script async src="https://maps.googleapis.com/maps/api/js?key=<%= api_key %>&callback=initMap"></script>
 
 </body>
 </html>
